Remove unnecessary casts and a few stray semi-colons
Casts probably date back to Java 1.4 and/or Spring 2

+++ core/src/main/java/org/osaf/cosmo/server/StatusSnapshot.java
@@ -30,10 +30,10 @@ public class StatusSnapshot {
     public StatusSnapshot() {
         Runtime runtime = Runtime.getRuntime();
 
-        totalMemory = (long) (runtime.totalMemory() / 1024);
-        freeMemory = (long) (runtime.freeMemory() / 1024);
-        usedMemory = (long) totalMemory - freeMemory;
-        maxMemory = (long) (runtime.maxMemory() / 1024);
+        totalMemory = runtime.totalMemory() / 1024;
+        freeMemory = runtime.freeMemory() / 1024;
+        usedMemory = totalMemory - freeMemory;
+        maxMemory = runtime.maxMemory() / 1024;
     }
 
     /**
